Moved some things from CommandSwerveDrivetrain.java to Drive.java. Reset pose now can be controlled more directly.
